added webpack, linter and dependencies
