+++ .eslintignore
@@ -1,0 +1,3 @@
+webpack.config.js
+node_modules
+build
