+++ .gitignore
@@ -1,0 +1,7 @@
+.vscode/
+.DS_Store
+package-lock.json
+node_modules/
+dist/
+build/
+.firebase/
