+++ src/javascripts/main.js
@@ -1,0 +1,9 @@
+import projects from './components/Projects/projects';
+
+import '../styles/main.scss';
+
+const init = () => {
+  projects.init();
+};
+
+init();
